# internal/server/api/openapi.yaml
openapi: "3.1.0"
info:
  title: express-botx API
  description: HTTP API for sending messages to eXpress corporate messenger via BotX API. Supports webhooks from Alertmanager and Grafana.
  version: dev
  license:
    name: MIT
    url: https://github.com/lavr/express-botx/blob/main/LICENSE
servers:
  - url: "{scheme}://{host}{basePath}"
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost:8080
      basePath:
        default: /api/v1
security:
  - bearerAuth: []
  - apiKeyHeader: []
paths:
...
  /alertmanager:
    post:
      operationId: alertmanager
      summary: Alertmanager webhook
      description: |
        Receives webhook notifications from Prometheus Alertmanager.
        Renders alerts through a Go template and sends the result to eXpress.

        **Multi-chat fan-out.** `?chat_id=` may list several chats separated by
        commas (`?chat_id=a,b,c`); the rendered alert is delivered best-effort to
        each. The response is always a `MultiSendResponse` (see the breaking-change
        note on `/send`), even for a single or default chat.

        Chat resolution priority:
        1. `?chat_id=` query parameter (comma-separated for fan-out)
        2. `default_chat_id` from webhook config
        3. Global default chat (`default: true` in chats section)
        4. Single chat alias from config
      parameters:
        - name: chat_id
          in: query
          description: Target chat UUID or alias; comma-separated for multi-chat fan-out (overrides config default)
          schema:
            type: string
        - name: bot
          in: query
          description: Bot name (required in multi-bot mode unless chat has a default bot binding)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertmanagerWebhook"
            example:
              version: "4"
              groupKey: '{}:{alertname="HighCPU"}'
              status: firing
              receiver: express
              groupLabels:
                alertname: HighCPU
              commonLabels:
                alertname: HighCPU
                env: prod
              commonAnnotations:
                summary: CPU is too high
              externalURL: http://alertmanager:9093
              alerts:
                - status: firing
                  labels:
                    alertname: HighCPU
                    severity: critical
                    instance: web-01:9090
                  annotations:
                    summary: "CPU > 90% on web-01"
                  startsAt: "2026-03-13T20:00:00.000Z"
                  endsAt: "0001-01-01T00:00:00Z"
                  generatorURL: http://prometheus:9090/graph
      responses:
        "200":
          description: |
            Alert delivered to at least one chat. Body is a `MultiSendResponse`
            (`results` with a per-chat `sync_id`; partial failures in `errors`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSendResponse"
        "400":
          description: Request-level error (malformed JSON, no alerts, empty chat_id, template error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No credentials provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Delivery to every listed chat failed (`MultiSendResponse` with `ok:false`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSendResponse"
...
  /grafana:
    post:
      operationId: grafana
      summary: Grafana webhook
      description: |
        Receives webhook notifications from Grafana Alerting.
        Renders alerts through a Go template and sends the result to eXpress.

        **Multi-chat fan-out.** `?chat_id=` may list several chats separated by
        commas (`?chat_id=a,b,c`); the rendered alert is delivered best-effort to
        each. The response is always a `MultiSendResponse` (see the breaking-change
        note on `/send`), even for a single or default chat.

        Chat resolution priority:
        1. `?chat_id=` query parameter (comma-separated for fan-out)
        2. `default_chat_id` from webhook config
        3. Global default chat (`default: true` in chats section)
        4. Single chat alias from config
      parameters:
        - name: chat_id
          in: query
          description: Target chat UUID or alias; comma-separated for multi-chat fan-out (overrides config default)
          schema:
            type: string
        - name: bot
          in: query
          description: Bot name (required in multi-bot mode unless chat has a default bot binding)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrafanaWebhook"
            example:
              receiver: express
              status: firing
              orgId: 1
              alerts:
                - status: firing
                  labels:
                    alertname: HighCPU
                    grafana_folder: Production
                  annotations:
                    summary: "CPU > 90% on web-01"
                  startsAt: "2026-03-13T20:00:00.000Z"
                  endsAt: "0001-01-01T00:00:00Z"
                  fingerprint: abc123def456
                  dashboardURL: http://grafana:3000/d/dashboard-uid
                  panelURL: http://grafana:3000/d/dashboard-uid?viewPanel=1
                  silenceURL: http://grafana:3000/alerting/silence/new
                  values:
                    B: 95.2
              groupLabels:
                alertname: HighCPU
              commonLabels:
                alertname: HighCPU
              commonAnnotations:
                summary: "CPU > 90% on web-01"
              externalURL: http://grafana:3000
              version: "1"
              groupKey: '{}:{alertname="HighCPU"}'
              truncatedAlerts: 0
              title: "[FIRING:1] HighCPU (Production)"
              state: alerting
              message: "**Firing**\n\nValue: B=95.2"
      responses:
        "200":
          description: |
            Alert delivered to at least one chat. Body is a `MultiSendResponse`
            (`results` with a per-chat `sync_id`; partial failures in `errors`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSendResponse"
        "400":
          description: Request-level error (malformed JSON, no alerts, empty chat_id, template error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No credentials provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Delivery to every listed chat failed (`MultiSendResponse` with `ok:false`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSendResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <api-key>"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: "X-API-Key: <api-key>"
    botSignature:
      type: apiKey
      in: header
      name: X-Bot-Signature
      description: HMAC-SHA256 signature for bot secret auth
    callbackJWT:
      type: http
      scheme: bearer
      description: "JWT signed with bot secret (HS256). Used for callback endpoints from eXpress platform."
    gitlabToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
      description: "X-Gitlab-Token: <secret token> — secret token configured on the GitLab webhook. Matched against the default `server.gitlab.secret` or any `server.gitlab.senders[].secret` (a sender token scopes delivery to that sender's chats)."
  schemas:
...
    AlertmanagerWebhook:
      type: object
      required: [status, alerts]
      properties:
        version:
          type: string
          example: "4"
        groupKey:
          type: string
        status:
          type: string
          enum: [firing, resolved]
        receiver:
          type: string
        groupLabels:
          type: object
          additionalProperties:
            type: string
        commonLabels:
          type: object
          additionalProperties:
            type: string
        commonAnnotations:
          type: object
          additionalProperties:
            type: string
        externalURL:
          type: string
          format: uri
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
    AlertItem:
      type: object
      properties:
        status:
          type: string
          enum: [firing, resolved]
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        generatorURL:
          type: string
          format: uri
    GrafanaWebhook:
      type: object
      required: [status, alerts]
      properties:
        receiver:
          type: string
        status:
          type: string
          enum: [firing, resolved]
        orgId:
          type: integer
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/GrafanaAlertItem"
        groupLabels:
          type: object
          additionalProperties:
            type: string
        commonLabels:
          type: object
          additionalProperties:
            type: string
        commonAnnotations:
          type: object
          additionalProperties:
            type: string
        externalURL:
          type: string
          format: uri
        version:
          type: string
        groupKey:
          type: string
        truncatedAlerts:
          type: integer
        title:
          type: string
          description: Pre-formatted alert title from Grafana
        state:
          type: string
          enum: [alerting, ok, no_data, pending]
        message:
          type: string
          description: Pre-formatted alert message from Grafana (markdown)
    GrafanaAlertItem:
      type: object
      properties:
        status:
          type: string
          enum: [firing, resolved]
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        generatorURL:
          type: string
          format: uri
        fingerprint:
          type: string
        silenceURL:
          type: string
          format: uri
        dashboardURL:
          type: string
          format: uri
        panelURL:
          type: string
          format: uri
        imageURL:
          type: string
          format: uri
        values:
          type: object
          additionalProperties: true
...
    MultiSendResponse:
      type: object
      description: |
        The uniform multi-chat fan-out response shared by `/send`, `/alertmanager`,
        `/grafana` and `/gitlab`. `ok` is true when at least one chat received the
        message (HTTP `200` sync / `202` async) and false when delivery to every
        chat failed (HTTP `502`). `results` lists each successful delivery and
        `errors` each failed one; on a partial failure both are present. Order
        follows the requested chats. Request-level failures (bad JSON, empty
        `chat_id`, invalid status, unsupported media) are NOT reported here — they
        keep the `{"ok":false,"error":"..."}` form with `400`/`415`.
      properties:
        ok:
          type: boolean
          example: true
        results:
          type: array
          description: Successful deliveries, one per chat.
          items:
            type: object
            properties:
              chat:
                type: string
                description: Target chat alias or UUID as requested.
                example: backend-mrs
              sync_id:
                type: string
                description: Synchronous delivery id (present for sync sends).
                example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              request_id:
                type: string
                description: Enqueue request id (present for async/enqueue sends).
                example: "0d6d7f87-0a2f-4c5b-b0d4-4d0b705a77e2"
              queued:
                type: boolean
                description: True when the message was queued (async/enqueue).
                example: true
        errors:
          type: array
          description: Failed deliveries, one per chat (chat/bot resolution or upstream send).
          items:
            type: object
            properties:
              chat:
                type: string
                example: frontend-mrs
              error:
                type: string
                example: "upstream error: 502 Bad Gateway"
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "chat_id is required"